Integrate the relationship store with mirage!!!!!!!
The server now has a relationships object, with defineOne, pushMany,
etc.

One small improvement (I think) is that a Response object was being
created around the route handler's unserialized response.
Then it was being unwrapped, serialized, and rewrapped in the
SerializerRegistry. I did discard the Registry temporarily, but it still
seemed like the wrong place to be doing that. I moved it to the route
handler instead, serializing the handler's return value before creating
the response code and wrapping it in the Response object.

Definitely discarded some functionality, and broke a lot more than that.
- Ripped out delegating to application or model-specific serializers,
  instead hardcoding my serializer.
- Discarded resolving promises returned in route handlers
- Almost definitely broke all the shorthand and object route handlers
- The serializers were taking the schema, but mine doesn't, so I got rid
  of that. Not sure what they needed the schema for, but mine might need
  it in the future for the same reason.
- I haven't tried just returning models in the route handler yet - I'm
  calling this.serialize().